refactor setUser mutation
